cleanup & warnings for modern gcc

+++ src/101b90_len_8f0.c
@@ -4,7 +4,7 @@
 #include "ld_addrs.h"
 
 #ifdef SHIFT
-#define SPRITE_ROM_START _1943000_ROM_START + 0x10
+#define SPRITE_ROM_START (u32) _1943000_ROM_START + 0x10
 #else
 #define SPRITE_ROM_START 0x1943000 + 0x10
 #endif
